Check to see if current user making updates is an admin before doing 'admin' operations during update
When using the users module in situations where a user could be, say updating their profile. The user is probably not an admin. Errors were being thrown after the update occurred because the the save routine doesn't check if the user making the modification to the user record has permissions to do things (make admin user, delete from admin users).

This modification adds an additional dependency upon "session" module which is used after the initial user save to verify if the current user has _admin permissions in order to elevate/demote user to system _admin.

Signed-off-by: Jim Klo <jim.klo@sri.com>